# src/main/resources/com/sonicle/webtop/tasks/openapi-v2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebTop Tasks",
    "contact": {
      "email": "dev-team@sonicle.com"
    },
    "version": "v2",
    "description": "This is Task service API enpoint."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "auth-basic": []
    },
    {
      "auth-bearer": []
    }
  ],
  "paths": {
...
    "/eas/folders/{folderId}/messages": {
      "post": {
        "tags": [
          "eas"
        ],
        "summary": "Adds a message",
        "description": "Creates new Task into specified Category.",
        "operationId": "addMessage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EasSyncTaskUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EasSyncTaskStat"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "body"
      },
      "parameters": [
        {
          "name": "folderId",
          "in": "path",
          "description": "Folder ID",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "1234"
        }
      ]
    },
    "/eas/folders/{folderId}/messages-stats": {
      "get": {
        "tags": [
          "eas"
        ],
        "summary": "Gets sync messages data for a folder",
        "description": "Returns sync informations for the specified Category.",
        "operationId": "getMessagesStats",
        "parameters": [
          {
            "name": "cutoffDate",
            "in": "query",
            "description": "Cut-off date (ISO date/time YYYYMMDD'T'HHMMSS'Z')",
            "schema": {
              "type": "string"
            },
            "example": "20150912T000000Z"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EasSyncTaskStat"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          }
        }
      },
      "parameters": [
        {
          "name": "folderId",
          "in": "path",
          "description": "Folder ID",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "1234"
        }
      ]
    },
    "/eas/folders/{folderId}/messages/{id}": {
      "get": {
        "tags": [
          "eas"
        ],
        "summary": "Get a single message",
        "description": "Gets the specified Task.",
        "operationId": "getMessage",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EasSyncTask"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Message not found",
            "content": {}
          }
        }
      },
      "put": {
        "tags": [
          "eas"
        ],
        "summary": "Updates a message",
        "description": "Updates the specified Task.",
        "operationId": "updateMessage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EasSyncTaskUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EasSyncTaskStat"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Message not found",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      },
      "delete": {
        "tags": [
          "eas"
        ],
        "summary": "Deletes a message",
        "description": "Deletes the specified Task.",
        "operationId": "deleteMessage",
        "responses": {
          "204": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Message not found",
            "content": {}
          }
        }
      },
      "parameters": [
        {
          "name": "folderId",
          "in": "path",
          "description": "Folder ID",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "1234"
        },
        {
          "name": "id",
          "in": "path",
          "description": "Message ID",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "1234"
        }
      ]
    },
...
  },
  "components": {
    "schemas": {
...
      "EasSyncTaskStat": {
        "required": [
          "etag",
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Task ID (internal)"
          },
          "etag": {
            "type": "string",
            "description": "Revision tag"
          }
        },
        "description": "Transport sync informations of a single task."
      },
      "EasSyncTask": {
        "required": [
          "etag",
          "id",
          "impo",
          "prvt",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Task ID (internal)"
          },
          "etag": {
            "type": "string",
            "description": "Revision tag"
          },
          "subject": {
            "type": "string",
            "description": "Subject"
          },
          "start": {
            "type": "string",
            "description": "Start date/time (ISO date/time YYYYMMDD’T’HHMMSS’Z’)"
          },
          "due": {
            "type": "string",
            "description": "Due date/time (ISO date/time YYYYMMDD’T’HHMMSS’Z’)"
          },
          "status": {
            "type": "string",
            "description": "Completion status"
          },
          "complOn": {
            "type": "string",
            "description": "Completed date/time (ISO date/time YYYYMMDD’T’HHMMSS’Z’)"
          },
          "impo": {
            "type": "integer",
            "description": "Priority flag (0=low, 1=normal, 2=high)"
          },
          "prvt": {
            "type": "boolean",
            "description": "Private flag"
          },
          "notes": {
            "type": "string",
            "description": "Description"
          }
        },
        "description": "Represent a single task."
      },
      "EasSyncTaskUpdate": {
        "title": "",
        "required": [
          "impo",
          "prvt"
        ],
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "description": "Subject"
          },
          "start": {
            "type": "string",
            "description": "Start date/time (ISO date/time YYYYMMDD'T'HHMMSS'Z')"
          },
          "due": {
            "type": "string",
            "description": "Due date/time (ISO date/time YYYYMMDD'T'HHMMSS'Z')"
          },
          "complOn": {
            "type": "string",
            "description": "Completed date/time (ISO date/time YYYYMMDD'T'HHMMSS'Z')"
          },
          "impo": {
            "type": "integer",
            "description": "Priority flag (0=low, 1=normal, 2=high)"
          },
          "prvt": {
            "type": "boolean",
            "description": "Private flag"
          },
          "notes": {
            "type": "string",
            "description": "Description"
          }
        },
        "description": "Represent task's updateable fields"
      },
...
    },
    "securitySchemes": {
      "auth-basic": {
        "type": "http",
        "scheme": "basic"
      },
      "auth-bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}